Update tests for createRoomWithCode function
